Split out intent functions into their own files. Also some general code clean up to get closer to PEP8 complience.
